chore: cleanup test suite and add code coverage
- Add pytest-cov to dev dependencies and configure coverage in pyproject.toml
- Update CI to generate coverage.json on main branch pushes
- Add dynamic coverage badge to README

Test suite cleanup (57 → 46 files, 405 → 384 tests, 0 coverage lost):
- DELETE 12 files: dead/duplicate/trivial tests
  - test_oxford_api.py (0 tests, legacy script)
  - test_gallica_resolver_unit.py (100% duplicate)
  - test_discovery_resolvers_unit.py (100% duplicate)
  - test_review_regressions.py (6 mock-wiring tests, redundant with test_http_client)
  - test_search_apis.py (tested removed function)
  - 6 micro-files (1-2 tests each) merged into parent test files
- MERGE useful tests from deleted files into canonical targets:
  - Resolver edge cases → test_discovery_resolvers.py
  - Route registration → test_studio_handlers.py
  - Mirador/viewer → test_studio_handlers.py
  - Layout filter → test_library_components.py
  - PDF generation → test_export_studio_pdf.py
  - Polling intervals → test_settings_handlers.py
- SIMPLIFY test_settings_ui_coverage.py (62 → 11 critical keys)
- REMOVE duplicate assertions (toast, export capabilities)
