+++ .github/workflows/ci.yml
@@ -77,8 +77,48 @@ jobs:
       run: |
         python -m pip install --upgrade pip
         if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
-        pip install pytest pytest-timeout
+        pip install pytest pytest-timeout pytest-cov
 
     - name: Run ${{ matrix.lane }} tests
       run: |
-        pytest -v --tb=short -m "${{ matrix.marker }}"
+        if [ "${{ matrix.lane }}" = "fast" ]; then
+          pytest -v --tb=short -m "${{ matrix.marker }}" --cov=src --cov-report=term-missing --cov-report=json:coverage.json
+        else
+          pytest -v --tb=short -m "${{ matrix.marker }}"
+        fi
+
+    - name: Upload coverage artifact
+      if: matrix.lane == 'fast'
+      uses: actions/upload-artifact@v4
+      with:
+        name: coverage-report
+        path: coverage.json
+
+  coverage-badge:
+    needs: test
+    if: github.event_name == 'push' && github.ref == 'refs/heads/main'
+    runs-on: ubuntu-latest
+    permissions:
+      contents: write
+    timeout-minutes: 5
+    steps:
+    - uses: actions/checkout@v4
+
+    - name: Download coverage artifact
+      uses: actions/download-artifact@v4
+      with:
+        name: coverage-report
+
+    - name: Update coverage.json and push
+      run: |
+        if [ -f coverage.json ]; then
+          git config user.name "github-actions[bot]"
+          git config user.email "41898282+github-actions[bot]@users.noreply.github.com"
+          git add coverage.json
+          if git diff --cached --quiet; then
+            echo "No coverage changes to commit"
+          else
+            git commit -m "chore: update coverage.json [skip ci]"
+            git push
+          fi
+        fi
