+++ requirements-dev.txt
@@ -9,3 +9,4 @@ pyright
 # Testing
 pytest
 pytest-timeout
+pytest-cov
